API: move 1.0.1 specific classes into its own package
