+++ src/main/java/org/semux/api/v1_0_1/response/CreateAccountResponse.java
@@ -4,21 +4,21 @@
  * Distributed under the MIT software license, see the accompanying file
  * LICENSE or https://opensource.org/licenses/mit-license.php
  */
-package org.semux.api.response;
+package org.semux.api.v1_0_1.response;
 
-import org.semux.api.ApiHandlerResponse;
+import org.semux.api.v1_0_1.ApiHandlerResponse;
 
 import com.fasterxml.jackson.annotation.JsonProperty;
 
-public class GetDelegateResponse extends ApiHandlerResponse {
+public class CreateAccountResponse extends ApiHandlerResponse {
 
     @JsonProperty("result")
-    public final Types.DelegateType delegate;
+    public final String accountAddress;
 
-    public GetDelegateResponse(
+    public CreateAccountResponse(
             @JsonProperty("success") Boolean success,
-            @JsonProperty("result") Types.DelegateType delegate) {
+            @JsonProperty("result") String accountAddress) {
         super(success, null);
-        this.delegate = delegate;
+        this.accountAddress = accountAddress;
     }
 }
