+++ src/main/java/org/semux/api/v1_0_1/response/GetBlockResponse.java
@@ -4,21 +4,21 @@
  * Distributed under the MIT software license, see the accompanying file
  * LICENSE or https://opensource.org/licenses/mit-license.php
  */
-package org.semux.api.response;
+package org.semux.api.v1_0_1.response;
 
-import org.semux.api.ApiHandlerResponse;
+import org.semux.api.v1_0_1.ApiHandlerResponse;
 
 import com.fasterxml.jackson.annotation.JsonProperty;
 
-public class GetRootResponse extends ApiHandlerResponse {
+public class GetBlockResponse extends ApiHandlerResponse {
 
     @JsonProperty("result")
-    public final String result;
+    public final Types.BlockType block;
 
-    public GetRootResponse(
+    public GetBlockResponse(
             @JsonProperty("success") Boolean success,
-            @JsonProperty("result") String result) {
+            @JsonProperty("result") Types.BlockType block) {
         super(success, null);
-        this.result = result;
+        this.block = block;
     }
 }
