+++ src/main/java/org/semux/api/v1_0_1/response/GetDelegateResponse.java
@@ -4,21 +4,21 @@
  * Distributed under the MIT software license, see the accompanying file
  * LICENSE or https://opensource.org/licenses/mit-license.php
  */
-package org.semux.api.response;
+package org.semux.api.v1_0_1.response;
 
-import org.semux.api.ApiHandlerResponse;
+import org.semux.api.v1_0_1.ApiHandlerResponse;
 
 import com.fasterxml.jackson.annotation.JsonProperty;
 
-public class GetLatestBlockNumberResponse extends ApiHandlerResponse {
+public class GetDelegateResponse extends ApiHandlerResponse {
 
     @JsonProperty("result")
-    public final Long blockNumber;
+    public final Types.DelegateType delegate;
 
-    public GetLatestBlockNumberResponse(
+    public GetDelegateResponse(
             @JsonProperty("success") Boolean success,
-            @JsonProperty("result") Long blockNumber) {
+            @JsonProperty("result") Types.DelegateType delegate) {
         super(success, null);
-        this.blockNumber = blockNumber;
+        this.delegate = delegate;
     }
 }
