+++ src/main/java/org/semux/api/v1_0_1/response/GetDelegatesResponse.java
@@ -4,23 +4,23 @@
  * Distributed under the MIT software license, see the accompanying file
  * LICENSE or https://opensource.org/licenses/mit-license.php
  */
-package org.semux.api.response;
+package org.semux.api.v1_0_1.response;
 
 import java.util.List;
 
-import org.semux.api.ApiHandlerResponse;
+import org.semux.api.v1_0_1.ApiHandlerResponse;
 
 import com.fasterxml.jackson.annotation.JsonProperty;
 
-public class ListAccountsResponse extends ApiHandlerResponse {
+public class GetDelegatesResponse extends ApiHandlerResponse {
 
     @JsonProperty("result")
-    public final List<String> accounts;
+    public final List<Types.DelegateType> delegates;
 
-    public ListAccountsResponse(
+    public GetDelegatesResponse(
             @JsonProperty("success") Boolean success,
-            @JsonProperty("result") List<String> accounts) {
+            @JsonProperty("result") List<Types.DelegateType> delegates) {
         super(success, null);
-        this.accounts = accounts;
+        this.delegates = delegates;
     }
 }
