+++ src/main/java/org/semux/api/v1_0_1/response/GetLatestBlockNumberResponse.java
@@ -4,23 +4,21 @@
  * Distributed under the MIT software license, see the accompanying file
  * LICENSE or https://opensource.org/licenses/mit-license.php
  */
-package org.semux.api.response;
+package org.semux.api.v1_0_1.response;
 
-import java.util.List;
-
-import org.semux.api.ApiHandlerResponse;
+import org.semux.api.v1_0_1.ApiHandlerResponse;
 
 import com.fasterxml.jackson.annotation.JsonProperty;
 
-public class GetValidatorsResponse extends ApiHandlerResponse {
+public class GetLatestBlockNumberResponse extends ApiHandlerResponse {
 
     @JsonProperty("result")
-    public final List<String> validators;
+    public final Long blockNumber;
 
-    public GetValidatorsResponse(
+    public GetLatestBlockNumberResponse(
             @JsonProperty("success") Boolean success,
-            @JsonProperty("result") List<String> validators) {
+            @JsonProperty("result") Long blockNumber) {
         super(success, null);
-        this.validators = validators;
+        this.blockNumber = blockNumber;
     }
 }
