+++ src/main/java/org/semux/api/v1_0_1/response/GetTransactionLimitsResponse.java
@@ -4,12 +4,12 @@
  * Distributed under the MIT software license, see the accompanying file
  * LICENSE or https://opensource.org/licenses/mit-license.php
  */
-package org.semux.api.response;
+package org.semux.api.v1_0_1.response;
 
 import static org.semux.core.TransactionType.DELEGATE;
 
 import org.semux.Kernel;
-import org.semux.api.ApiHandlerResponse;
+import org.semux.api.v1_0_1.ApiHandlerResponse;
 import org.semux.core.TransactionType;
 
 import com.fasterxml.jackson.annotation.JsonCreator;
