+++ src/main/java/org/semux/api/v1_0_1/response/GetTransactionResponse.java
@@ -4,21 +4,21 @@
  * Distributed under the MIT software license, see the accompanying file
  * LICENSE or https://opensource.org/licenses/mit-license.php
  */
-package org.semux.api.response;
+package org.semux.api.v1_0_1.response;
 
-import org.semux.api.ApiHandlerResponse;
+import org.semux.api.v1_0_1.ApiHandlerResponse;
 
 import com.fasterxml.jackson.annotation.JsonProperty;
 
-public class GetAccountResponse extends ApiHandlerResponse {
+public class GetTransactionResponse extends ApiHandlerResponse {
 
     @JsonProperty("result")
-    public final Types.AccountType account;
+    public final Types.TransactionType transaction;
 
-    public GetAccountResponse(
+    public GetTransactionResponse(
             @JsonProperty("success") Boolean success,
-            @JsonProperty("result") Types.AccountType account) {
+            @JsonProperty("result") Types.TransactionType transaction) {
         super(success, null);
-        this.account = account;
+        this.transaction = transaction;
     }
 }
