+++ src/main/java/org/semux/api/v1_0_1/response/GetVoteResponse.java
@@ -4,21 +4,21 @@
  * Distributed under the MIT software license, see the accompanying file
  * LICENSE or https://opensource.org/licenses/mit-license.php
  */
-package org.semux.api.response;
+package org.semux.api.v1_0_1.response;
 
-import org.semux.api.ApiHandlerResponse;
+import org.semux.api.v1_0_1.ApiHandlerResponse;
 
 import com.fasterxml.jackson.annotation.JsonProperty;
 
-public class GetTransactionResponse extends ApiHandlerResponse {
+public class GetVoteResponse extends ApiHandlerResponse {
 
     @JsonProperty("result")
-    public final Types.TransactionType transaction;
+    public final Long vote;
 
-    public GetTransactionResponse(
+    public GetVoteResponse(
             @JsonProperty("success") Boolean success,
-            @JsonProperty("result") Types.TransactionType transaction) {
+            @JsonProperty("result") Long vote) {
         super(success, null);
-        this.transaction = transaction;
+        this.vote = vote;
     }
 }
