+++ src/main/java/org/semux/api/v1_0_1/response/GetVotesResponse.java
@@ -4,21 +4,23 @@
  * Distributed under the MIT software license, see the accompanying file
  * LICENSE or https://opensource.org/licenses/mit-license.php
  */
-package org.semux.api.response;
+package org.semux.api.v1_0_1.response;
 
-import org.semux.api.ApiHandlerResponse;
+import java.util.Map;
+
+import org.semux.api.v1_0_1.ApiHandlerResponse;
 
 import com.fasterxml.jackson.annotation.JsonProperty;
 
-public class CreateAccountResponse extends ApiHandlerResponse {
+public class GetVotesResponse extends ApiHandlerResponse {
 
     @JsonProperty("result")
-    public final String accountAddress;
+    public final Map<String, Long> votes;
 
-    public CreateAccountResponse(
+    public GetVotesResponse(
             @JsonProperty("success") Boolean success,
-            @JsonProperty("result") String accountAddress) {
+            @JsonProperty("result") Map<String, Long> votes) {
         super(success, null);
-        this.accountAddress = accountAddress;
+        this.votes = votes;
     }
 }
