+++ src/main/java/org/semux/api/v1_0_1/response/Types.java
@@ -4,7 +4,7 @@
  * Distributed under the MIT software license, see the accompanying file
  * LICENSE or https://opensource.org/licenses/mit-license.php
  */
-package org.semux.api.response;
+package org.semux.api.v1_0_1.response;
 
 import java.util.List;
 import java.util.stream.Collectors;
